refactor: package and code overhaul
