--- src/main/java/com/rocs/infirmary/desktop/application/app/facade/low/stock/medicine/LowStockMedicineFacade.java
@@ -1,8 +1,0 @@
-package com.rocs.infirmary.desktop.application.app.facade.low.stock.medicine;
-
-import com.rocs.infirmary.desktop.application.model.inventory.LowStockItem;
-
-import java.util.List;
-public interface LowStockMedicineFacade {
-    List<LowStockItem> findAllLowStockMedicine();
-}
