--- src/main/java/com/rocs/infirmary/desktop/application/data/dao/medical/record/StudentMedicalRecordDao.java
@@ -1,9 +1,0 @@
-package com.rocs.infirmary.desktop.application.data.dao.medical.record;
-
-
-import com.rocs.infirmary.desktop.application.model.medical.record.StudentMedicalRecord;
-
-public interface StudentMedicalRecordDao {
-    StudentMedicalRecord getMedicalInformationByLRN(long LRN);
-
-}
