--- src/main/java/com/rocs/infirmary/desktop/application/model/inventory/LowStockItem.java
@@ -1,19 +1,0 @@
-package com.rocs.infirmary.desktop.application.model.inventory;
-
-public class LowStockItem {
-    private String description;
-    private int quantityAvailable;
-
-    public LowStockItem(String description, int quantityAvailable) {
-        this.description = description;
-        this.quantityAvailable = quantityAvailable;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public int getQuantityAvailable() {
-        return quantityAvailable;
-    }
-}
