--- src/main/java/com/rocs/infirmary/desktop/application/model/person/Person.java
@@ -1,128 +1,0 @@
-package com.rocs.infirmary.desktop.application.model.person;
-
-import java.util.Date;
-
-public class Person {
-    private long id;
-    private String firstName;
-    private String middleName;
-    private String lastName;
-    private int age;
-    private Date birthdate;
-    private String gender;
-    private String email;
-    private String address;
-    private int contactNumber;
-
-    public Person(long id, String firstName, String middleName, String lastName, int age, Date birthdate, String gender, String email, String address, int contactNumber) {
-        this.id = id;
-        this.firstName = firstName;
-        this.middleName = middleName;
-        this.lastName = lastName;
-        this.age = age;
-        this.birthdate = birthdate;
-        this.gender = gender;
-        this.email = email;
-        this.address = address;
-        this.contactNumber = contactNumber;
-    }
-
-    public Person() {
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getMiddleName() {
-        return middleName;
-    }
-
-    public void setMiddleName(String middleName) {
-        this.middleName = middleName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public Date getBirthdate() {
-        return birthdate;
-    }
-
-    public void setBirthdate(Date birthdate) {
-        this.birthdate = birthdate;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public int getContactNumber() {
-        return contactNumber;
-    }
-
-    public void setContactNumber(int contactNumber) {
-        this.contactNumber = contactNumber;
-    }
-
-    @Override
-    public String toString() {
-        return "Person{" +
-                "id=" + id +
-                ", firstName='" + firstName + '\'' +
-                ", middleName='" + middleName + '\'' +
-                ", lastName='" + lastName + '\'' +
-                ", age=" + age +
-                ", birthdate=" + birthdate +
-                ", gender='" + gender + '\'' +
-                ", email='" + email + '\'' +
-                ", address='" + address + '\'' +
-                ", contactNumber=" + contactNumber +
-                '}';
-    }
-}
