+++ src/main/java/com/rocs/infirmary/desktop/data/dao/student/record/StudentMedicalRecordDao.java
@@ -1,0 +1,9 @@
+package com.rocs.infirmary.desktop.data.dao.student.record;
+
+
+import com.rocs.infirmary.desktop.data.model.person.student.Student;
+
+public interface StudentMedicalRecordDao {
+    Student getMedicalInformationByLRN(long LRN);
+
+}
