+++ src/main/java/com/rocs/infirmary/desktop/data/model/inventory/Inventory.java
@@ -1,0 +1,32 @@
+package com.rocs.infirmary.desktop.data.model.inventory;
+
+public class Inventory {
+    private int inventoryId;
+    private String description;
+    private int quantityAvailable;
+
+
+    public int getInventoryId() {
+        return inventoryId;
+    }
+
+    public void setInventoryId(int inventoryId) {
+        this.inventoryId = inventoryId;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public int getQuantityAvailable() {
+        return quantityAvailable;
+    }
+
+    public void setQuantityAvailable(int quantityAvailable) {
+        this.quantityAvailable = quantityAvailable;
+    }
+}
